Fix label detection on long labels
My printer (a Citizen CLP-521) starts in metric mode, and is set to metric mode anyway.
This commit fixes two places where lengths were being sent in hundredths of inches, when tenths of millimeters is correct.

This fixes an issue I was experiencing where the second label would print, but the printer raised an error that it could not find the third.

It is untested on other hardware